+++ app/Transformer/Api/Mastodon/v1/StatusTransformer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Transformer\Api;
+namespace App\Transformer\Api\Mastodon\v1;
 
 use App\Status;
 use League\Fractal;
@@ -11,6 +11,8 @@ class StatusTransformer extends Fractal\TransformerAbstract
     protected $defaultIncludes = [
         'account',
         'media_attachments',
+        'mentions',
+        'tags',
     ];
 
     public function transform(Status $status)
@@ -23,31 +25,27 @@ class StatusTransformer extends Fractal\TransformerAbstract
             'in_reply_to_account_id'    => $status->in_reply_to_profile_id,
             'reblog'                    => null,
             'content'                   => $status->rendered ?? $status->caption,
-            'created_at'                => $status->created_at->format('c'),
+            'created_at'                => $status->created_at->toJSON(),
             'emojis'                    => [],
+            'replies_count'             => 0,
             'reblogs_count'             => $status->reblogs_count != 0 ? $status->reblogs_count: $status->shares()->count(),
             'favourites_count'          => $status->likes_count != 0 ? $status->likes_count: $status->likes()->count(),
-            'reblogged'                 => $status->shared(),
-            'favourited'                => $status->liked(),
+            'reblogged'                 => null,
+            'favourited'                => null,
             'muted'                     => null,
             'sensitive'                 => (bool) $status->is_nsfw,
             'spoiler_text'              => $status->cw_summary ?? '',
             'visibility'                => $status->visibility ?? $status->scope,
+            'mentions'                  => [],
+            'tags'                      => [],
+            'card'                      => null,
+            'poll'                      => null,
             'application'               => [
                 'name'      => 'web',
                 'website'   => null
              ],
             'language'                  => null,
             'pinned'                    => null,
-            'mentions'                  => [],
-            'tags'                      => [],
-            'pf_type'                   => $status->type ?? $status->setType(),
-            'reply_count'               => (int) $status->reply_count,
-            'comments_disabled'         => $status->comments_disabled ? true : false,
-            'thread'                    => false,
-            'replies'                   => [],
-            'parent'                    => [],
-            'place'                     => $status->place
         ];
     }
 
@@ -60,11 +58,25 @@ class StatusTransformer extends Fractal\TransformerAbstract
 
     public function includeMediaAttachments(Status $status)
     {
-        return Cache::remember('status:transformer:media:attachments:'.$status->id, now()->addDays(14), function() use($status) {
+        return Cache::remember('mastoapi:status:transformer:media:attachments:'.$status->id, now()->addDays(14), function() use($status) {
             if(in_array($status->type, ['photo', 'video', 'photo:album', 'loop', 'photo:video:album'])) {
                 $media = $status->media()->orderBy('order')->get();
                 return $this->collection($media, new MediaTransformer());
             }
         });
+    }
+
+    public function includeMentions(Status $status)
+    {
+        $mentions = $status->mentions;
+
+        return $this->collection($mentions, new MentionTransformer());
+    }
+
+    public function includeTags(Status $status)
+    {
+        $hashtags = $status->hashtags;
+
+        return $this->collection($hashtags, new HashtagTransformer());
     }
 }