+++ resources/views/settings/applications.blade.php
@@ -16,9 +16,4 @@
 
 @push('scripts')
 <script type="text/javascript" src="{{mix('js/developers.js')}}"></script>
-<script type="text/javascript">
-	new Vue({ 
-		el: '#content' 
-	});
-</script>
 @endpush